Implement a more natural Left Hand Mode
Moving right or left moves the player slightly forward all the time due
to the hexagonal layout. This perfectly fits the Q and E keys which are
next to the W key which is the key for moving forward. Also, moving left
and right is something that the player probably won't use as often as
turning left and right. So by putting the turning on the AD of the WASD
keys that players are used to, they immediately are comfortable with the
controls.